Move Insomnia to Postman to a separate file
